updated getCaseType to take tenant from X-Auth-Rolls

+++ src/main/java/uk/gov/digital/ho/hocs/info/entities/Sla.java
@@ -5,6 +5,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.util.spi.LocaleNameProvider;
 
 @Entity
 @Table(name = "sla")
@@ -24,6 +25,6 @@ public class Sla {
 
     @Getter
     @Column(name = "value")
-    private int value;
+    private Long value;
 
 }
